修复bug

+++ src/main/java/com/ruoyi/project/storage/service/impl/BackendBackendCustomerServiceImpl.java
@@ -75,13 +75,13 @@ public class BackendBackendCustomerServiceImpl implements BackendCustomerService
         final Map<String, Object> byUsername = findCustomerByUsername(customerVO.getUserName());
         final Map<String, Object> byEmail = findCustomerByEmail(customerVO.getEmail());
         final Map<String, Object> byPhoneNumber = findCustomerByPhoneNumber(customerVO.getPhonenumber());
-        if ((Integer) byUsername.get("count") > 0) {
+        if ((Long) byUsername.get("count") > 0) {
             throw new CustomException("新增'" + customerVO.getUserName() + "'失败，登录账号已存在");
         }
-        if ((Integer) byEmail.get("count") > 0) {
+        if ((Long) byEmail.get("count") > 0) {
             throw new CustomException("新增'" + customerVO.getUserName() + "'失败，邮箱账号已存在");
         }
-        if ((Integer) byPhoneNumber.get("count") > 0) {
+        if ((Long) byPhoneNumber.get("count") > 0) {
             throw new CustomException("新增'" + customerVO.getUserName() + "'失败，手机号已存在");
         }
         checkCustomer(customerVO);
